Revert "settings: literal instead of casting"
This reverts commit 95e992a87c898f3e882ffdb415bb0ef9f80f613f.

+++ src/core/settings.h
@@ -444,7 +444,7 @@ struct Values {
     Setting<bool> renderer_debug;
     Setting<int> vulkan_device;
 
-    Setting<u16> resolution_factor = Setting(1U);
+    Setting<u16> resolution_factor = Setting(static_cast<u16>(1));
     Setting<int> aspect_ratio;
     Setting<int> max_anisotropy;
     Setting<bool> use_frame_limit;
